Added explicit conversions to remove some warnings.

+++ Reduction/src/Game.cpp
@@ -227,8 +227,8 @@ void Game::update()
 	if (m_WallScale >= 0.3 - WALL_SPEED)
 	{
 		m_WallScale += WALL_SPEED;
-		m_WallRect.w = m_OriginalWallWidth * m_WallScale;
-		m_WallRect.h = m_OriginalWallHeight * m_WallScale;
+		m_WallRect.w = (int) (m_OriginalWallWidth * m_WallScale);
+		m_WallRect.h = (int) (m_OriginalWallHeight * m_WallScale);
 	}
 
 }
